app/requirements.txt
numpy~=1.26.0
torch~=2.2.1
fastapi~=0.110.0
uvicorn~=0.27.1
setuptools~=69.1.1
mlflow-skinny~=2.11.1
python-dotenv~=1.0.1
python-multipart~=0.0.9
aiofiles~=23.2.1
